README
# instagramBot
A bot bot that accomplishes simple tasks on Instagram.

My objective was to create a bot that can log into user accounts on Instagram, 
travels to chosen "hashtags" and begins liking and commenting pictures
and following the accounts of some of these pictures.

The goal was to passively gain followers, likes and comments on my own personal 
instagram as a side-effect of the "engagement" with other users.

There were a few small hurdles that I had to work around during this project:

1. I had to do my best to avoid bot-like behavior that could be flagged by instagram.
From the limited research I have done, I learned that actions performed with the same 
time intervals raises suspicion. I made sure that my program would sleep at different
random intervals between actions.

2. I had to avoid exceeding Instagram's maximum permitted actions per hour.
Accounts that Have been active for over a year are allowed to safely like around 150 
pictures per hour without being restricted. At first I tested this theory and had the 
bot like pictures at a far faster rate and recieved a 1 hour "timeout" from liking 
pictures because of the spamminess of my bot.

3. I had to avoid leaving the same comment over and over, which would immediately look 
like spam and potentially get my account banned.
To avoid this, I made a list of 20 different comments that are vague and general purpose
and have the bot leave a comment on 1 out of every 10 pictures (on average) using 
one of the 20 random comments.

To my delight, I actually recieved a lot of user engagement due to the activity of the bot.
I gained several hundred followers in just a couples weeks and recieved replies to many
of the bot's comments. It seemed to pass the "real-human" user test.